Fix segmentation fault
Strcpy caused some unexpected behaviour depending on the actual length
of the str string.

Strlen returns the number of characters between the beginning of the
string and the terminating null character whereas strcpy null-terminates
all copied strings.

Allocating one additional byte for the null terminator fixed the issue for me.